Multiple levels first step
I changed the ghost speed in each level in the inspector so that each
level gets progressively harder, and I based the percentage increase on
what is found in the Pacman dossier. I did the same thing for Pacman
himself.

In order to do the same for the speed of the frightened ghosts I
substituted the constant by which their regular speed was multiplied
for a public float variable, which I then altered appropriately for
each scene in the inspector.

I tried to change to create versions of the pac dot scripts for each
level, so that the winning condition for each scene would be to load
the next one. However, the pacdot objects are prefabs, and therefore my
plan to apply a different script to them in each scene failed. I’m
still trying to figure out how else to do this.